requirements.txt
torch
torchvision
streamlit
pillow
kagglehub
matplotlib
seaborn
scikit-learn
numpy
tensorflow
keras